refactor: move matchers and node helpers to separate files
